Specs added that check existence of branches in a MR
When creating or updating a merge request via API, `source_branch` and
`target_branch` can be set to names that do not reference existing
branches (see issue #2915). The UI does not allow to select non-existing
branches, but the API currently does and does not validate the branch names.
It is therefore advised to respond with an appropriate status code (422 comes
to mind) if a branch does not exist. A few tests are added that describe the
desired behavior for these cases.