API: project creation allows namespace id, also access to namespaces
Currently there is no good way to see which namespaces a user is allowed to create projects
in. It is now possible to create a project with a valid namespace id, e.g. as administrator
to create a project in the global namespace or a user is allowed to create a project within
a group.

Tests added for access of namespaces via API.